Move getLocalFunctionWrapperCode logic
